backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
xlrd>=2.0.1
python-multipart
sqlalchemy
pyodbc
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-dotenv
azure-storage-blob
aiohttp
pydantic-settings
email-validator
xlsxwriter
reportlab>=4.0.0
